+++ packages/language-service/ivy/test/legacy/language_service_adapter_spec.ts
@@ -8,7 +8,7 @@
 
 import * as ts from 'typescript/lib/tsserverlibrary';
 
-import {LanguageServiceAdapter} from '../language_service_adapter';
+import {LanguageServiceAdapter} from '../../language_service_adapter';
 
 import {MockService, setup, TEST_TEMPLATE} from './mock_host';
